회원가입 화면 / 시작하기 버튼 화면 비율 깨짐 관련 수정.

+++ Assets/Scripts/Stores/User.cs
@@ -518,9 +518,9 @@ public class record {
     public int user;
     public int count;
     public string riding_time;
-    public int total_distance;
-    public int max_speed;
-    public int avg_speed;
+    public float total_distance;
+    public float max_speed;
+    public float avg_speed;
 }
 
 [System.Serializable]
